Add files via upload
HOW THE SYSTEM WORKS

On application launch, a login window is generated requesting the user to sign in. If the user does not have an account, the user is required to register. There is a registration window that deals with the creation of new users. The login and the registration windows are connected to the database server and synchronize with data from the user table. The concept behind the registration window is MYSQL INSERT query while the concept behind the login window is MYSQL SELECT query. If on login the user credential matches a record in the database then the user is redirected to the home page.
The home page is made up of a menu in which the user can choose an option such as viewing record, adding record, deleting record, updating record, querying database as well as generating report. All the system functionalities are based on MYSQL queries. For example, to view record we use MYSQL SELECT query and to delete record we use MYSQL DELETE query. Moving around the windows in the app is handled with JavaFX via FXML loading. Also transition from one window to another is made smooth with the help of JavaFX animation built in classes.
The tracking sheets (report) are generated by Jasper Report Library. It is an open source reporting engine which provides the ability to deliver rich content onto a printer, a screen, or into various formats such as PDF, HTML, XLS, RTF, ODT, CSV, TXT and XML files.